LayoutTests/fast/repaint/position-relative-container-with-abs-pos-child-expected.txt

PASS window.internals.repaintRectsAsText().indexOf('0 0 50 50') is not -1
PASS successfullyParsed is true

TEST COMPLETE
